requirements.txt
autopep8==1.5
bcrypt==3.1.7
beautifulsoup4==4.9.0
billboard.py==6.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dominate==2.5.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
pycparser==2.20
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
